Start using classes instead of functions to generate widgets
