+++ lib/routes/study_pages/flashcard_summary.dart
@@ -215,7 +215,7 @@ class FlashcardSummary extends StatelessWidget {
                   flashcardData: snapshot.data!,
                 );
               } else {
-                return LoadingScreen.wholeScreen(context);
+                return LoadingScreen();
               }
             },
           ),
