+++ lib/widgets/custom_snack_bars.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
-class ScreenMessageSnackBar extends SnackBar {
-  ScreenMessageSnackBar.onError({super.key, required String text})
+class CustomSnackBars extends SnackBar {
+  CustomSnackBars.onError({super.key, required String text})
       : super(content: Text(text), duration: Duration(seconds: 4));
-  ScreenMessageSnackBar.onSuccess({super.key, required String text})
+  CustomSnackBars.onSuccess({super.key, required String text})
       : super(content: Text(text), duration: Duration(seconds: 4));
 
   void show(BuildContext context) {
