+++ lib/widgets/loading_screen.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:loading_animation_widget/loading_animation_widget.dart';
 
-class LoadingScreen {
-  static Widget wholeScreen(BuildContext context) {
+class LoadingScreen extends StatelessWidget {
+  const LoadingScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
     return Center(
       child: LoadingAnimationWidget.horizontalRotatingDots(color: Theme.of(context).colorScheme.primary, size: 150.0),
     );
